test(lsp): organize protocol trace tests
Move the test-only router helper and protocol tracing harness into the router test modules. Split notification suppression and in-flight trace-level transitions into focused cases without changing server behavior.
